Project/model/requirements.txt
flask
tensorflow-cpu==2.13.0
opencv-python
numpy
gunicorn
requests
python-dotenv
